+++ enterprise/frontend/src/embedding-sdk/components/public/InteractiveQuestion/InteractiveQuestion.tsx
@@ -1,19 +1,33 @@
-import { useMemo } from "react";
+import { type PropsWithChildren, useMemo } from "react";
 
 import { InteractiveQuestionResult } from "embedding-sdk/components/private/InteractiveQuestionResult";
 import { withPublicComponentWrapper } from "embedding-sdk/components/private/PublicComponentWrapper";
-import { InteractiveQuestionProvider } from "embedding-sdk/components/public/InteractiveQuestion/context";
 import type { SdkClickActionPluginsConfig } from "embedding-sdk/lib/plugins";
 import type { CardId } from "metabase-types/api";
 
-interface InteractiveQuestionProps {
+import {
+  BackButton,
+  FilterBar,
+  QuestionResetButton,
+  Title,
+  Filter,
+  FilterButton,
+  Summarize,
+  SummarizeButton,
+  Notebook,
+  NotebookButton,
+  QuestionVisualization,
+} from "./components";
+import { InteractiveQuestionProvider } from "./context";
+
+type InteractiveQuestionProps = PropsWithChildren<{
   questionId: CardId;
   withResetButton?: boolean;
   withTitle?: boolean;
   customTitle?: React.ReactNode;
   plugins?: SdkClickActionPluginsConfig;
   height?: string | number;
-}
+}>;
 
 export const _InteractiveQuestion = ({
   questionId,
@@ -22,6 +36,7 @@ export const _InteractiveQuestion = ({
   customTitle,
   plugins,
   height,
+  children = null,
 }: InteractiveQuestionProps): JSX.Element | null => {
   const { location, params } = useMemo(
     () => getQuestionParameters(questionId),
@@ -33,17 +48,18 @@ export const _InteractiveQuestion = ({
       location={location}
       params={params}
       componentPlugins={plugins}
-      customTitle={customTitle}
-      withResetButton={withResetButton}
-      withTitle={withTitle}
     >
-      <InteractiveQuestionResult height={height} />
+      {children ?? (
+        <InteractiveQuestionResult
+          height={height}
+          customTitle={customTitle}
+          withResetButton={withResetButton}
+          withTitle={withTitle}
+        />
+      )}
     </InteractiveQuestionProvider>
   );
 };
-
-export const InteractiveQuestion =
-  withPublicComponentWrapper(_InteractiveQuestion);
 
 export const getQuestionParameters = (questionId: CardId) => {
   return {
@@ -57,3 +73,33 @@ export const getQuestionParameters = (questionId: CardId) => {
     },
   };
 };
+
+const InteractiveQuestion = withPublicComponentWrapper(
+  _InteractiveQuestion,
+) as typeof _InteractiveQuestion & {
+  BackButton: typeof BackButton;
+  FilterBar: typeof FilterBar;
+  Filter: typeof Filter;
+  FilterButton: typeof FilterButton;
+  ResetButton: typeof QuestionResetButton;
+  Title: typeof Title;
+  Summarize: typeof Summarize;
+  SummarizeButton: typeof SummarizeButton;
+  Notebook: typeof Notebook;
+  NotebookButton: typeof NotebookButton;
+  QuestionVisualization: typeof QuestionVisualization;
+};
+
+InteractiveQuestion.BackButton = BackButton;
+InteractiveQuestion.FilterBar = FilterBar;
+InteractiveQuestion.Filter = Filter;
+InteractiveQuestion.FilterButton = FilterButton;
+InteractiveQuestion.ResetButton = QuestionResetButton;
+InteractiveQuestion.Title = Title;
+InteractiveQuestion.Summarize = Summarize;
+InteractiveQuestion.SummarizeButton = SummarizeButton;
+InteractiveQuestion.Notebook = Notebook;
+InteractiveQuestion.NotebookButton = NotebookButton;
+InteractiveQuestion.QuestionVisualization = QuestionVisualization;
+
+export { InteractiveQuestion };